# Improve SELECTs performance and add `RowCursor::{decoded_bytes,received_bytes}`