--- a/benches/common.rs
+++ b/benches/common.rs
@@ -1,6 +1,14 @@
 #![allow(dead_code)] // typical for common test/bench modules :(
 
-use std::{convert::Infallible, future::Future, net::SocketAddr, thread};
+use std::{
+    convert::Infallible,
+    future::Future,
+    net::SocketAddr,
+    pin::Pin,
+    sync::atomic::{AtomicU32, Ordering},
+    thread,
+    time::Duration,
+};
 
 use bytes::Bytes;
 use futures::stream::StreamExt;
@@ -11,7 +19,13 @@ use hyper::{
     service, Request, Response,
 };
 use hyper_util::rt::{TokioIo, TokioTimer};
-use tokio::{net::TcpListener, runtime};
+use tokio::{
+    net::TcpListener,
+    runtime,
+    sync::{mpsc, oneshot},
+};
+
+use clickhouse::error::Result;
 
 pub(crate) struct ServerHandle;
 
@@ -38,14 +52,7 @@ where
         }
     };
 
-    thread::spawn(move || {
-        runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(serving);
-    });
-
+    run_on_st_runtime("server", serving);
     ServerHandle
 }
 
@@ -56,4 +63,63 @@ pub(crate) async fn skip_incoming(request: Request<Incoming>) {
     while let Some(result) = body.next().await {
         result.unwrap();
     }
+}
+
+pub(crate) struct RunnerHandle {
+    tx: mpsc::UnboundedSender<Run>,
+}
+
+struct Run {
+    future: Pin<Box<dyn Future<Output = Result<Duration>> + Send>>,
+    callback: oneshot::Sender<Result<Duration>>,
+}
+
+impl RunnerHandle {
+    pub(crate) fn run(
+        &self,
+        f: impl Future<Output = Result<Duration>> + Send + 'static,
+    ) -> Duration {
+        let (tx, rx) = oneshot::channel();
+
+        self.tx
+            .send(Run {
+                future: Box::pin(f),
+                callback: tx,
+            })
+            .unwrap();
+
+        rx.blocking_recv().unwrap().unwrap()
+    }
+}
+
+pub(crate) fn start_runner() -> RunnerHandle {
+    let (tx, mut rx) = mpsc::unbounded_channel::<Run>();
+
+    run_on_st_runtime("testee", async move {
+        while let Some(run) = rx.recv().await {
+            let result = run.future.await;
+            let _ = run.callback.send(result);
+        }
+    });
+
+    RunnerHandle { tx }
+}
+
+fn run_on_st_runtime(name: &str, f: impl Future + Send + 'static) {
+    let name = name.to_string();
+    thread::Builder::new()
+        .name(name.clone())
+        .spawn(move || {
+            let no = AtomicU32::new(0);
+            runtime::Builder::new_current_thread()
+                .enable_all()
+                .thread_name_fn(move || {
+                    let no = no.fetch_add(1, Ordering::Relaxed);
+                    format!("{name}-{no}")
+                })
+                .build()
+                .unwrap()
+                .block_on(f);
+        })
+        .unwrap();
 }