--- a/benches/select_numbers.rs
+++ b/benches/select_numbers.rs
@@ -9,13 +9,14 @@ struct Data {
 
 async fn bench(name: &str, compression: Compression) {
     let start = std::time::Instant::now();
-    let sum = tokio::spawn(do_bench(compression)).await.unwrap();
+    let (sum, dec_mbytes, rec_mbytes) = tokio::spawn(do_bench(compression)).await.unwrap();
     assert_eq!(sum, 124999999750000000);
     let elapsed = start.elapsed();
-    println!("{name:>8}  {elapsed:>7.3?}");
+    let throughput = dec_mbytes / elapsed.as_secs_f64();
+    println!("{name:>8}  {elapsed:>7.3?}  {throughput:>4.0} MiB/s  {rec_mbytes:>4.0} MiB");
 }
 
-async fn do_bench(compression: Compression) -> u64 {
+async fn do_bench(compression: Compression) -> (u64, f64, f64) {
     let client = Client::default()
         .with_compression(compression)
         .with_url("http://localhost:8123");
@@ -30,12 +31,16 @@ async fn do_bench(compression: Compression) -> u64 {
         sum += row.no;
     }
 
-    sum
+    let dec_bytes = cursor.decoded_bytes();
+    let dec_mbytes = dec_bytes as f64 / 1024.0 / 1024.0;
+    let recv_bytes = cursor.received_bytes();
+    let recv_mbytes = recv_bytes as f64 / 1024.0 / 1024.0;
+    (sum, dec_mbytes, recv_mbytes)
 }
 
 #[tokio::main]
 async fn main() {
-    println!("compress  elapsed");
+    println!("compress  elapsed  throughput  received");
     bench("none", Compression::None).await;
     #[cfg(feature = "lz4")]
     bench("lz4", Compression::Lz4).await;