--- a/src/compression/lz4.rs
+++ b/src/compression/lz4.rs
@@ -9,37 +9,38 @@ use futures::{ready, stream::Stream};
 use lz4_flex::block;
 
 use crate::{
-    buflist::BufList,
+    bytes_ext::BytesExt,
     error::{Error, Result},
+    response::Chunk,
 };
 
 const MAX_COMPRESSED_SIZE: u32 = 1024 * 1024 * 1024;
 
 pub(crate) struct Lz4Decoder<S> {
     stream: S,
-    chunks: BufList<Bytes>,
+    bytes: BytesExt,
     meta: Option<Lz4Meta>,
-    buffer: Vec<u8>,
 }
 
 impl<S> Stream for Lz4Decoder<S>
 where
     S: Stream<Item = Result<Bytes>> + Unpin,
 {
-    type Item = Result<Bytes>;
+    type Item = Result<Chunk>;
 
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         let meta = loop {
-            let size = self.chunks.remaining();
-            let required_size = self.meta.as_ref().map_or(LZ4_META_SIZE, |m| {
-                m.compressed_size as usize - LZ4_HEADER_SIZE
-            });
+            let size = self.bytes.remaining();
+            let required_size = self
+                .meta
+                .as_ref()
+                .map_or(LZ4_META_SIZE, Lz4Meta::total_size);
 
             if size < required_size {
                 let stream = Pin::new(&mut self.stream);
                 match ready!(stream.poll_next(cx)) {
                     Some(Ok(chunk)) => {
-                        self.chunks.push(chunk);
+                        self.bytes.extend(chunk);
                         continue;
                     }
                     Some(Err(err)) => return Some(Err(err)).into(),
@@ -51,7 +52,7 @@ where
                 }
             }
 
-            assert!(size >= required_size);
+            debug_assert!(size >= required_size);
 
             match self.meta.take() {
                 Some(meta) => break meta,
@@ -59,16 +60,18 @@ where
             };
         };
 
-        let bytes = self.read_data(meta)?;
-        self.chunks.commit();
-        Poll::Ready(Some(Ok(bytes)))
+        let data = self.read_data(&meta)?;
+        let net_size = meta.total_size();
+        self.bytes.advance(net_size);
+
+        Poll::Ready(Some(Ok(Chunk { data, net_size })))
     }
 }
 
 // Meta = checksum + header
 // - [16b] checksum
 // - [ 1b] magic number (0x82)
-// - [ 4b] compressed size
+// - [ 4b] compressed size (data + header)
 // - [ 4b] uncompressed size
 const LZ4_CHECKSUM_SIZE: usize = 16;
 const LZ4_HEADER_SIZE: usize = 9;
@@ -82,11 +85,15 @@ struct Lz4Meta {
 }
 
 impl Lz4Meta {
-    fn read(mut buffer: impl Buf) -> Result<Lz4Meta> {
-        let checksum = buffer.get_u128_le();
-        let magic = buffer.get_u8();
-        let compressed_size = buffer.get_u32_le();
-        let uncompressed_size = buffer.get_u32_le();
+    fn total_size(&self) -> usize {
+        LZ4_CHECKSUM_SIZE + self.compressed_size as usize
+    }
+
+    fn read(mut bytes: &[u8]) -> Result<Lz4Meta> {
+        let checksum = bytes.get_u128_le();
+        let magic = bytes.get_u8();
+        let compressed_size = bytes.get_u32_le();
+        let uncompressed_size = bytes.get_u32_le();
 
         if magic != LZ4_MAGIC {
             return Err(Error::Decompression("incorrect magic number".into()));
@@ -103,11 +110,11 @@ impl Lz4Meta {
         })
     }
 
-    fn write_checksum(&self, mut buffer: impl BufMut) {
+    fn write_checksum(&self, mut buffer: &mut [u8]) {
         buffer.put_u128_le(self.checksum);
     }
 
-    fn write_header(&self, mut buffer: impl BufMut) {
+    fn write_header(&self, mut buffer: &mut [u8]) {
         buffer.put_u8(LZ4_MAGIC);
         buffer.put_u32_le(self.compressed_size);
         buffer.put_u32_le(self.uncompressed_size);
@@ -118,36 +125,28 @@ impl<S> Lz4Decoder<S> {
     pub(crate) fn new(stream: S) -> Self {
         Self {
             stream,
-            chunks: BufList::default(),
+            bytes: BytesExt::default(),
             meta: None,
-            buffer: Vec::new(),
         }
     }
 
     fn read_meta(&mut self) -> Result<Lz4Meta> {
-        assert!(self.chunks.remaining() >= LZ4_META_SIZE);
-        Lz4Meta::read(&mut self.chunks)
+        Lz4Meta::read(self.bytes.slice())
     }
 
-    fn read_data(&mut self, meta: Lz4Meta) -> Result<Bytes> {
-        assert!(self.chunks.remaining() >= meta.compressed_size as usize - LZ4_HEADER_SIZE);
+    fn read_data(&mut self, meta: &Lz4Meta) -> Result<Bytes> {
+        let total_size = meta.total_size();
+        let bytes = &self.bytes.slice()[..total_size];
 
-        self.buffer.resize(meta.compressed_size as usize, 0);
-        meta.write_header(&mut self.buffer[..]);
-
-        // TODO: if we have a whole frame in one chunk, extra copying can be avoided.
-        self.chunks
-            .copy_to_slice(&mut self.buffer[LZ4_HEADER_SIZE..]);
-
-        let actual_checksum = calc_checksum(&self.buffer);
+        let actual_checksum = calc_checksum(&bytes[LZ4_CHECKSUM_SIZE..]);
         if actual_checksum != meta.checksum {
             return Err(Error::Decompression("checksum mismatch".into()));
         }
 
-        let mut uncompressed = vec![0u8; meta.uncompressed_size as usize];
-        let len = decompress(&self.buffer[LZ4_HEADER_SIZE..], &mut uncompressed)?;
-        debug_assert_eq!(len as u32, meta.uncompressed_size);
+        let uncompressed = block::decompress_size_prepended(&bytes[(LZ4_META_SIZE - 4)..])
+            .map_err(|err| Error::Decompression(err.into()))?;
 
+        debug_assert_eq!(uncompressed.len() as u32, meta.uncompressed_size);
         Ok(uncompressed.into())
     }
 }
@@ -155,10 +154,6 @@ impl<S> Lz4Decoder<S> {
 fn calc_checksum(buffer: &[u8]) -> u128 {
     let hash = cityhash_102_128(buffer);
     hash.rotate_right(64)
-}
-
-fn decompress(compressed: &[u8], uncompressed: &mut [u8]) -> Result<usize> {
-    block::decompress_into(compressed, uncompressed).map_err(|err| Error::Decompression(err.into()))
 }
 
 pub(crate) fn compress(uncompressed: &[u8]) -> Result<Bytes> {
@@ -195,9 +190,12 @@ async fn it_decompresses() {
     ];
 
     let source = vec![
-        245_u8, 5, 222, 235, 225, 158, 59, 108, 225, 31, 65, 215, 66, 66, 36, 92, 130, 34, 0, 0, 0,
-        23, 0, 0, 0, 240, 8, 1, 0, 2, 255, 255, 255, 255, 0, 1, 1, 1, 115, 6, 83, 116, 114, 105,
-        110, 103, 3, 97, 98, 99,
+        245_u8, 5, 222, 235, 225, 158, 59, 108, 225, 31, 65, 215, 66, 66, 36, 92,   // checksum
+        0x82, // magic number
+        34, 0, 0, 0, // compressed size (data + header)
+        23, 0, 0, 0, // uncompressed size
+        240, 8, 1, 0, 2, 255, 255, 255, 255, 0, 1, 1, 1, 115, 6, 83, 116, 114, 105, 110, 103, 3,
+        97, 98, 99,
     ];
 
     async fn test(chunks: &[&[u8]], expected: &[u8]) {
@@ -209,8 +207,12 @@ async fn it_decompresses() {
                 .collect::<Vec<_>>(),
         );
         let mut decoder = Lz4Decoder::new(stream);
-        let actual = decoder.try_next().await.unwrap();
-        assert_eq!(actual.as_deref(), Some(expected));
+        let actual = decoder.try_next().await.unwrap().unwrap();
+        assert_eq!(actual.data, expected);
+        assert_eq!(
+            actual.net_size,
+            chunks.iter().map(|s| s.len()).sum::<usize>()
+        );
     }
 
     // 1 chunk.