--- a/src/error.rs
+++ b/src/error.rs
@@ -38,12 +38,6 @@ pub enum Error {
     BadResponse(String),
     #[error("timeout expired")]
     TimedOut,
-
-    // Internally handled errors, not part of public API.
-    // XXX: move to another error?
-    #[error("internal error: too small buffer, need another {0} bytes")]
-    #[doc(hidden)]
-    TooSmallBuffer(usize),
 }
 
 assert_impl_all!(Error: StdError, Send, Sync);