--- a/tests/it/main.rs
+++ b/tests/it/main.rs
@@ -56,6 +56,7 @@ async fn flush_query_log(client: &Client) {
 
 mod compression;
 mod cursor_error;
+mod cursor_stats;
 mod insert;
 mod inserter;
 mod ip;