## Summary

This PR introduces `RowCursor::{decoded_bytes,received_bytes}` methods to provide stats useful for telemetry on the client side.

Besides this, the performance of `SELECT` queries is significantly improved:
* `BufList` (to handle situations when one row is related to several HTTP chunks) is replaced with simple `Bytes` and cloning in slow path cases. So,  `RowBinaryDeserializer` now accepts `&[u8]`, which is much cheaper than `BufList`. This change achieves the most performance gain (~80%).
* Remove a complex logic with a closure in `RawCursor` (~30%).
* Expose `RowBinaryCursor` directly as `RowCursor` to avoid one extra state machine in the `RowCursor::next()` future (~5%).
* Remove extra cloning of buffers in the lz4 decoder (~7% in the `select/lz4` bench).
* Use [decompress_size_prepended](https://docs.rs/lz4_flex/latest/lz4_flex/block/fn.decompress_size_prepended.html) to avoid extra `vec![0; size]` if `lz4_flex` is not compiled with the `safe-decode` feature (~5% in the `select/lz4` bench).
* Don't copy bytes for borrowed usage (not covered by benchmarks).

Also, this is the first step in exposing `RawCursor` (https://github.com/ClickHouse/clickhouse-rs/pull/152). `RawCursor` will be exposed with the added `async collect(self) -> Result<Bytes>` method and examples in the following PRs.

### "select" bench
#### Intel(R) Xeon(R) Gold 6226R @ 2.90 GHz

Before (on 7b599ef82ecbd3aee5abc2edd3297cef7e25055b):
```
select/no compression   time:   [33.485 ns 33.565 ns 33.649 ns]
                        thrpt:  [680.19 MiB/s 681.90 MiB/s 683.53 MiB/s]
select/lz4              time:   [37.391 ns 37.448 ns 37.509 ns]
                        thrpt:  [610.20 MiB/s 611.19 MiB/s 612.13 MiB/s]
```

After:
```
select/no compression   time:   [14.051 ns 14.096 ns 14.146 ns]
                        thrpt:  [1.5801 GiB/s 1.5857 GiB/s 1.5908 GiB/s]
                 change:
                        time:   [-57.932% -57.576% -57.199%] (p = 0.00 < 0.05)
                        thrpt:  [+133.64% +135.71% +137.71%]
                        Performance has improved.
select/lz4              time:   [17.116 ns 17.182 ns 17.253 ns]
                        thrpt:  [1.2955 GiB/s 1.3009 GiB/s 1.3059 GiB/s]
                 change:
                        time:   [-54.684% -54.240% -53.801%] (p = 0.00 < 0.05)
                        thrpt:  [+116.46% +118.53% +120.67%]
                        Performance has improved.
```

#### AMD Ryzen 7 4800HS @ 2.90 GHz
Before (on 7b599ef82ecbd3aee5abc2edd3297cef7e25055b):
```
select/no compression   time:   [37.901 ns 37.981 ns 38.076 ns]
                        thrpt:  [601.12 MiB/s 602.63 MiB/s 603.90 MiB/s]
select/lz4              time:   [41.244 ns 41.318 ns 41.401 ns]
                        thrpt:  [552.85 MiB/s 553.95 MiB/s 554.95 MiB/s]
```

After:
```
select/no compression   time:   [15.214 ns 15.288 ns 15.368 ns]
                        thrpt:  [1.4545 GiB/s 1.4621 GiB/s 1.4691 GiB/s]
                 change:
                        time:   [-60.124% -59.455% -58.770%] (p = 0.00 < 0.05)
                        thrpt:  [+142.54% +146.64% +150.78%]
                        Performance has improved.
select/lz4              time:   [17.214 ns 17.279 ns 17.353 ns]
                        thrpt:  [1.2880 GiB/s 1.2936 GiB/s 1.2984 GiB/s]
                 change:
                        time:   [-58.850% -58.060% -57.372%] (p = 0.00 < 0.05)
                        thrpt:  [+134.59% +138.43% +143.01%]
                        Performance has improved.
```

### "select_numbers" bench
CPU: AMD Ryzen 7 4800HS @ 2.90 GHz
CH: 24.9.2 (localhost, docker)

Before (on 7b599ef82ecbd3aee5abc2edd3297cef7e25055b):
```
compress  elapsed
    none   6.257s
     lz4   8.301s
```
(chosen worst among five runs)

After:
```
compress  elapsed  throughput  received
    none   2.868s  1330 MiB/s  3815 MiB
     lz4   6.301s   605 MiB/s  1908 MiB
```
(chosen worst among five runs)

## Checklist
Delete items not relevant to your PR:
- [x] Unit and integration tests covering the common scenarios were added
- [x] A human-readable description of the changes was provided so that we can include it in CHANGELOG later
- [ ] For significant changes, documentation in README and https://github.com/ClickHouse/clickhouse-docs was updated with further explanations or tutorials