@Arjentix 2024-10-18
I've tested this branch on my query.

Row structure looks something like this:
```rust
pub struct Row {
    pub id: u32,
    pub string_1: String,
    pub string_2: String,
    pub names: Vec<String>,
    pub kinds: Vec<String>,
    pub templates: Vec<String>,
    pub states: Vec<String>,
    pub some_stuff_names: Vec<Vec<String>>,
    pub some_stuff_contents: Vec<Vec<bytes::Bytes>>,
}
```

Most of the time it worked well even before this branch. But then data became quite bigger and it started getting too slow. Here are the details for a query like `SELECT * FROM table WHERE id in [12345]` which loads a lot of data.
I compare `fix/mock-with-advanced-time` (I think there is no difference with master, so I'll call it master) and `feat/query_stats`

master: ~37s
feat/query_stats: ~8s
clickhouse client: ~6.5s

So this is much closer to the native client time but still there is a room for optimizations